Update brother-chan-troi-duc-niem--i-have-a-dream.md
edited this section per Br Duc N's request
I have a dream that one day many different spiritual and faith traditions, humanitarian groups and NGOs will come together to co-create and follow a code of Global Ethics similar to the Five Mindfulness Trainings. I am convinced that this can serve as a connecting element between different groups and engaged people, and as a foundation, and inspiration for a new movement.